Adds feature for explicitly vendoring libuv
This change introduces new feature `vendored-cmake` that allows users to
explicitly opt into building and vendoring libuv at compile time.
This is useful in cases where you want the final application to be
better isolated from system dependencies.

The existing behavior (searching for the library via pkg-config and then
building it if not found) is left as the default.

This introduces an optional dependency on cmake.